fix: ทำcontroller มาใช้ในroute
