+++ backend-demo-test/test/db/user.spec.js
@@ -1,0 +1,9 @@
+const { expect } = require("chai");
+const mongoose = require("mongoose");
+
+describe("users", () => {
+    before(() => mongoose.connect(process.env.MONGO_URL));
+
+    it('find all users is')
+
+})
